+++ util/database.py
@@ -1,3 +1,16 @@
+from pathlib import Path
 from flask_sqlalchemy import SQLAlchemy
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
 
 db = SQLAlchemy()
+
+
+def create_session():
+    parent_dir = Path(__file__).parent.parent.absolute()
+    dbPath = str(parent_dir) + '/mybuzzn.db'
+    engine = create_engine('sqlite:///%s' % dbPath)
+    Session = sessionmaker(bind=engine)
+    session = Session()
+    return session
